+++ packages/core-js/modules/esnext.typed-array.from-async.js
@@ -1,4 +1,5 @@
 'use strict';
+var getBuiltIn = require('../internals/get-built-in');
 var aConstructor = require('../internals/a-constructor');
 var arrayFromAsync = require('../internals/array-from-async');
 var TYPED_ARRAYS_CONSTRUCTORS_REQUIRES_WRAPPERS = require('../internals/typed-array-constructors-require-wrappers');
@@ -12,8 +13,14 @@ var exportTypedArrayStaticMethod = ArrayBufferViewCore.exportTypedArrayStaticMet
 // https://github.com/tc39/proposal-array-from-async
 // eslint-disable-next-line -- required for .length
 exportTypedArrayStaticMethod('fromAsync', function fromAsync(asyncItems /* , mapfn = undefined, thisArg = undefined */) {
-  var C = aConstructor(this);
-  return arrayFromAsync.apply(Array, arguments).then(function (list) {
+  var C = this;
+  var argumentsLength = arguments.length;
+  var mapfn = argumentsLength > 1 ? arguments[1] : undefined;
+  var thisArg = argumentsLength > 2 ? arguments[2] : undefined;
+  return new (getBuiltIn('Promise'))(function (resolve) {
+    aConstructor(C);
+    resolve(arrayFromAsync(asyncItems, mapfn, thisArg));
+  }).then(function (list) {
     return arrayFromConstructorAndList(aTypedArrayConstructor(C), list);
   });
 }, TYPED_ARRAYS_CONSTRUCTORS_REQUIRES_WRAPPERS);